sdk: move agent interface in the SDK
Follow the Go best practices: define interfaces in the first package expecting
them. So the Agent interface needs to be defined in `sdk`, not `agent/types`.
